modules/cloud-function/vriddhi/login/src/requirements.txt
google-cloud-secret-manager==2.10.0
SQLAlchemy==1.3.12
boto3==1.26.44